change from int to double (cpp)

+++ studyHelper/timer.cpp
@@ -8,9 +8,9 @@ void Timer::start(int minutes)
     endTime = startTime + minutes;
 }
 
-int Timer::getTimeLeft()
+double Timer::getTimeLeft()
 {
-    int timeleft = endTime - ((millis() / 1000) * 60);
+    double timeleft = endTime - ((millis() / 1000) * 60);
     if (timeleft > 0)
         return timeleft;
     else
@@ -19,7 +19,7 @@ int Timer::getTimeLeft()
 
 bool Timer::isTimerDone()
 {
-    int currentTime = (millis() / 1000) * 60;
+    double currentTime = (millis() / 1000) * 60;
     if (currentTime >= endTime)
     {
       return true;   
@@ -29,6 +29,6 @@ bool Timer::isTimerDone()
 
 void Timer::restart()
 {
-    startTime = 0;
-    endTime = 0;
+    startTime = 0.0;
+    endTime = 0.0;
 }
